+++ test/unit/test_fracture_intersect_boundary.py
@@ -12,12 +12,10 @@ import unittest
 import numpy as np
 
 import porepy as pp
-from porepy.fracs.fractures import Fracture
-
 
 class TestFractureBoundaryIntersection(unittest.TestCase):
     def setup(self):
-        self.f_1 = Fracture(
+        self.f_1 = pp.Fracture(
             np.array([[0, 1, 1, 0], [0.5, 0.5, 0.5, 0.5], [0, 0, 1, 1]]),
             check_convexity=False,
         )
@@ -92,7 +90,7 @@ class TestFractureBoundaryIntersection(unittest.TestCase):
     def test_full_incline(self):
         self.setup()
         p = np.array([[-0.5, 0.5, 0.5, -0.5], [0.5, 0.5, 1.5, 1.5], [-0.5, -0.5, 1, 1]])
-        f = Fracture(p, check_convexity=False)
+        f = pp.Fracture(p, check_convexity=False)
         network = pp.FractureNetwork3d([f])
         network.impose_external_boundary(self.domain)
         p_known = np.array(
@@ -104,5 +102,4 @@ class TestFractureBoundaryIntersection(unittest.TestCase):
 
 
 if __name__ == "__main__":
-    TestFractureBoundaryIntersection().test_incline_in_plane()
     unittest.main()
